+++ packages/polaris-viz-core/src/utilities/index.ts
@@ -21,4 +21,3 @@ export {getValueFromXScale} from './getValueFromXScale';
 export {isValueWithinDomain} from './isValueWithinDomain';
 export {getClosestDivisibleNumber} from './getClosestDivisibleNumber';
 export {roundToDecimals} from './roundToDecimals';
-export {getAriaLabel} from './getAriaLabel';